TextChanged creado y otro cambios.

+++ Parcial1-AP1/Entidades/Evaluacion.cs
@@ -12,7 +12,7 @@ namespace Parcial1_AP1.Entidades
         private int IDEvaluacion;
         private DateTime fecha;
         private string Estudiante;
-        private int valor;
+        private decimal valor;
         private decimal logadro;
         private decimal perdido;
         private int pronostico;
@@ -26,7 +26,7 @@ namespace Parcial1_AP1.Entidades
         public int IDEvaluacion1 { get ; set; }
         public DateTime Fecha { get ; set ; }
         public string Estudiante1 { get ; set ; }
-        public int Valor { get ; set ; }
+        public decimal Valor { get ; set ; }
         public decimal Logadro { get ; set; }
         public decimal Perdido { get ; set ; }
         public int Pronostico { get ; set; }
